Fix FRI shutdown.
Now that KRCCommThread is not keeping the communication with FRI alive,
the delay between exiting the main loop and calling `stopFRI()`, which
consists on a call to `spinner.stop()`, is enough for the KRC to detect
a degraded communication quality and abruptly stop the robot.

Moving `spinner.stop()` to after `stopFRI()` makes it block. This is likely
due to some race condition with the controller spawners trying to stop and
unload controllers and blocking since controller manager is no longer updated.
Removed the `spinner.stop()` altogether since the process is shutting down
anyways.